requirements.txt
flask>=3.0.3
pdfplumber>=0.11.0
pypdf2>=3.0.0
pandas>=2.2.2
tabulate>=0.9.0
python-dateutil>=2.9.0
rich>=13.7.1
pytesseract>=0.3.13
Pillow>=10.4.0
gunicorn>=21.2.0
